+++ src/lib/prelude.ts
@@ -173,3 +173,16 @@ export type GetKeyOr<T, Key, Or> = Key extends keyof T ? T[Key] : Or
 import type { ConditionalSimplifyDeep } from 'type-fest/source/conditional-simplify.js'
 
 export type SimplifyDeep<T> = ConditionalSimplifyDeep<T, Function | Iterable<unknown> | Date, object>
+
+import fs from 'node:fs/promises'
+
+export const fileExists = async (path: string) => {
+  return Boolean(
+    await fs.stat(path).catch((_: unknown) => {
+      const error = errorFromMaybeError(_)
+      return `code` in error && typeof error.code === `string` && error.code === `ENOENT`
+        ? null
+        : Promise.reject(error)
+    }),
+  )
+}
